@@ -35,10 +35,10 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   crypto: ^3.0.1
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.7.0
   cached_network_image: ^3.2.0
   expandable: ^5.0.1
-  flutter_launcher_icons: ^0.9.2
+  flutter_launcher_icons: ^0.13.1
   # flutter_launcher_icons: ^0.9.2 
 
   # flutter_icons:
@@ -57,7 +57,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^1.0.0      
+  flutter_lints: ^3.0.2      
     
 
 # For information on the generic Dart part of this file, see the
