@@ -3,16 +3,18 @@ import 'package:flutter/material.dart';
 
 
 void main() {
-  runApp( MyApp());
+  runApp( const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return  MaterialApp(    
       
-      home: IntroPage(),
-       theme: ThemeData.dark(),
+      home: const IntroPage(),
+      theme: ThemeData.dark(),
       // darkTheme: ThemeData(colorScheme: ColorScheme.dark(onPrimary: Colors.black87)),
         themeMode: ThemeMode.light,
     );
